Issues-66: Moved RecyclerCollapsingToolbarActivity.kt to separate module.
